[upstream] Bug fixes (box2d #897)
rename null joint to filter joint (less confusing)
fix for polygon collision in the speculative region
example of kinematic target
fixed cast result consistency
gate demo
more character mover testing

box2d #875
box2d #880
box2d #868

+++ src/Box2D.NET.Samples/Samples/Benchmarks/BenchmarkCast.cs
@@ -128,7 +128,7 @@ public class BenchmarkCast : Sample
                         box = b2MakeBox(halfWidth, ratio * halfWidth);
                     }
 
-                    int category = (int)RandomIntRange(0, 2);
+                    int category = RandomIntRange(0, 2);
                     shapeDef.filter.categoryBits = (ulong)(1 << category);
                     if (category == 0)
                     {
